+++ BulbChipSTM32C071FBPx/Core/Inc/device/rgb_led.h
@@ -19,8 +19,8 @@ typedef struct RGBLed {
 	void (*startLedTimers)();
 	void (*stopLedTimers)();
 
-	uint16_t ms;
-	uint16_t msOfColorChange;
+	uint32_t ms;
+	uint32_t msOfColorChange;
 	bool showingTransientStatus;
 	uint8_t userRed;
 	uint8_t userGreen;
@@ -29,7 +29,7 @@ typedef struct RGBLed {
 
 bool rgbInit(RGBLed *device, RGBWritePwm *writeFn, uint16_t period, void (*startLedTimers)(), void (*stopLedTimers)());
 
-void rgbTask(RGBLed *device, uint16_t ms);
+void rgbTask(RGBLed *device, uint32_t ms);
 void rgbShowNoColor(RGBLed *device);
 void rgbShowUserColor(RGBLed *device, uint8_t red, uint8_t green, uint8_t blue);
 void rgbShowSuccess(RGBLed *device);
